Move `ProductWithQuantityStepperViewModel` to its own file. Update `ProductWithQuantityStepperView` preview.
